MemberRepository 조회 오류 해결

+++ jjikeobang-server/src/main/java/com/jjikeobang/member/repository/MemberRepository.java
@@ -15,7 +15,7 @@ public interface MemberRepository {
 
     void putMember(JoinMemberDTO member);
 
-    Member findById(long memberId);
+    Member findById(int memberId);
 
     Member findByLoginId(String userId);
 }
